pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = "A Python ToDoList application"
authors = ["Rose <rose729729@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = ">=1.1.1,<2.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
pytest = "^8.0.0"

[tool.poetry.scripts]
todolist = "app.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"